Added RecipeQuickView on map markers

+++ CulinaryMap/Models/Request/UpdateRecipeModel.cs
@@ -4,8 +4,8 @@
     {
         public int Id { get; set; }
         public int RegionId { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
         public string Name { get; set; }
         public string PrepMode { get; set; }
         public string? History { get; set; }
