+++ CulinaryMap/Models/Response/RecipeResponse.cs
@@ -8,8 +8,8 @@ namespace CulinaryMap.Models.Response
         public int Id { get; set; }
         public int RegionId { get; set; }
         public string Region { get; set; }
-        public int Latitude { get; set; }
-        public int Longitude { get; set; }
+        public decimal Latitude { get; set; }
+        public decimal Longitude { get; set; }
         public string Name { get; set; }
         public string PrepMode { get; set; }
         public string? History { get; set; }
